styled

+++ client/src/components/Authentication/Register/Register.js
@@ -1,11 +1,16 @@
 import React, {useState} from 'react'
 
+import '../Form.scss'
+
+import { toast } from 'react-toastify';
+
+
 export default function Register() {
 
   const [input,
-    setInput] = useState({name: "", email: "", password: ""})
+    setInput] = useState({name: "", email: "", password: "", confirmPassword: ""})
 
-  const {name, email, password} = input
+  const {name, email, password, confirmPassword} = input
 
   const onChange = (e) => {
     setInput({
@@ -16,40 +21,46 @@ export default function Register() {
 
   const onFormSubmit = async(e) => {
     e.preventDefault()
-    try {
+    if (password !== confirmPassword){
+      toast.error("Passwords don't match")
+    } else {
+      try {
 
-      const body = {
-        name,
-        email,
-        password
-      };
-
-      const baseURL = process.env.NODE_ENV === 'production' ? "api/auth/register" : "http://localhost:5000/api/auth/register"
-
-      const res = await fetch(baseURL, {
-          method: "POST",
-          headers: {
-              "Content-type": "application/json"
-          },
-          body: JSON.stringify(body)
-      })
-
-      const parseRes = await res.json()
-
-      if (parseRes.token) {
-        localStorage.setItem("token", parseRes.token)
-      } else {
-          console.log("Sheeeeeeeeesh")
+        const body = {
+          name,
+          email,
+          password
+        };
+  
+        const baseURL = process.env.NODE_ENV === 'production' ? "api/auth/register" : "http://localhost:5000/api/auth/register"
+  
+        const res = await fetch(baseURL, {
+            method: "POST",
+            headers: {
+                "Content-type": "application/json"
+            },
+            body: JSON.stringify(body)
+        })
+  
+        const parseRes = await res.json()
+  
+        if (parseRes.token) {
+          localStorage.setItem("token", parseRes.token)
+          toast.success("Registered Successfully")
+        } else {
+            console.log("Sheeeeeeeeesh")
+        }
+  
+      } catch (error) {
+        console.log(error.message)
       }
-
-    } catch (error) {
-      console.log(error.message)
-    }
+    } 
   }
 
   return (
     <div>
       <form onSubmit={onFormSubmit}>
+        <h1>Register</h1>
         <label>Username</label>
         <input type="text" name="name" value={name} onChange={e => onChange(e)}/>
         <label>Email</label>
@@ -59,6 +70,12 @@ export default function Register() {
           type="password"
           name="password"
           value={password}
+          onChange={e => onChange(e)}/>
+        <label>Confirm Password</label>
+        <input
+          type="password"
+          name="confirmPassword"
+          value={confirmPassword}
           onChange={e => onChange(e)}/>
         <button onClick={onFormSubmit}>Submit</button>
       </form>
